Jupyter: R Script for FRNT50 Calculations
Jupyter Notebook file with R Script for FRNT50 Calculations.  Script takes a folder of CSV files, reads in the data from each file, combines into a data.frame, calculates FRNT50 values using LL.3, and exports the results.  Current CSV format, which could be updated, provides setup and metadata in the first row, followed by the number of FFU in each well from an 8x12 grid (96 well plate).  The data is processed assuming that the plate has 7 dilutions and a undiluted/base dilution positive control in the 8th row, and is split into two groups, with 6 columns to a group.